pyproject.toml
[project]
name = "pp8k"
version = "0.1.0"
description = "Python driver for the Polaroid ProPalette 8000 film recorder"
readme = "README.md"
license = "LGPL-3.0-or-later"
requires-python = ">=3.10"
dependencies = ["Pillow"]

[project.scripts]
pp8k = "pp8k.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
